feat : move ResponseResource to common module
